# backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Middle Monitor API",
    "version": "1.0.0",
    "description": "REST API of the Middle Monitor observability platform.\n\nAuthentication is a bearer token: a personal API token (Settings, API tokens)\nor an organization API key. Agent install tokens are a separate credential and\nare only accepted on the /agents endpoints, through the X-Install-Token header.\n\nList endpoints accept limit and offset and return the total in X-Total-Count.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.middlemonitor.io",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal token or organization API key"
      },
      "installToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Install-Token",
        "description": "Agent install token, accepted only on the agent endpoints"
      }
    },
...
    "schemas": {
      "AgentScrapeConfig": {
        "type": "object",
        "properties": {
          "scrape_config": {
            "type": "string",
            "description": "YAML scrape fragment: a list of scrape targets, or a targets block. Empty clears it."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/v1/agents/config": {
      "get": {
        "operationId": "handleAgentConfig",
        "summary": "Agent config",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentScrapeConfig"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "Called by the agent itself, with its install token, to fetch its own scrape fragment.",
        "security": []
      }
    },
    "/api/v1/agents/download/install": {
      "get": {
        "operationId": "handleAgentInstallScript",
        "summary": "Agent install script",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      }
    },
    "/api/v1/agents/download/update": {
      "get": {
        "operationId": "handleAgentUpdateScript",
        "summary": "Agent update script",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      }
    },
    "/api/v1/agents/download/{os}/{arch}": {
      "get": {
        "operationId": "handleAgentDownload",
        "summary": "Agent download",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "Download the agent binary. Answers 304 to a conditional request (ETag is the sha256).",
        "parameters": [
          {
            "name": "os",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      }
    },
    "/api/v1/agents/download/{os}/{arch}/sha256": {
      "get": {
        "operationId": "handleAgentChecksum",
        "summary": "Agent checksum",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "The sha256 of one binary, in the shasum format.",
        "parameters": [
          {
            "name": "os",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      }
    },
    "/api/v1/agents/download/{version}/{os}/{arch}": {
      "get": {
        "operationId": "handleAgentDownload",
        "summary": "Agent download",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "Download a pinned version. 404 when that version is no longer the one on offer.",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "os",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": []
      }
    },
    "/api/v1/agents/latest": {
      "get": {
        "operationId": "handleAgentLatest",
        "summary": "Agent latest",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRelease"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "Version, sha256 and pinnable download URL of the agent binaries on offer.",
        "security": []
      }
    },
...
    "/api/v1/agents/register": {
      "post": {
        "operationId": "handleAgentRegister",
        "summary": "Agent register",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/2fa/disable": {
      "post": {
        "operationId": "handleDisable2FA",
        "summary": "Disable2 fa",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/2fa/setup": {
      "post": {
        "operationId": "handleSetup2FA",
        "summary": "Setup2 fa",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/2fa/verify": {
      "post": {
        "operationId": "handleEnable2FA",
        "summary": "Enable2 fa",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/accept-invite": {
      "post": {
        "operationId": "handleAcceptInvitation",
        "summary": "Accept invitation",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/api-tokens": {
      "get": {
        "operationId": "handleGetPersonalTokens",
        "summary": "Get personal tokens",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      },
      "post": {
        "operationId": "middleware.RequireMFAEnrollment",
        "summary": "Middleware.require mfaenrollment",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/api-tokens/{id}": {
      "delete": {
        "operationId": "handleDeletePersonalToken",
        "summary": "Delete personal token",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "security": []
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "handleForgotPassword",
        "summary": "Forgot password",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "handleOpenAPI",
        "summary": "Open api",
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "security": []
      }
    },
...
  }
}